feat: add loading tent component and integrate into various pages for improved loading experience
