+++ src/app/(reservation-flow)/loading.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import LoadingTent from '@/components/common/loading-tent';
+
+export default function Loading() {
+	return <LoadingTent />;
+}
