+++ src/components/pages/landing/facilities-section.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Button } from '@/components/ui/button';
 import { ArrowLeft, ArrowRight, ArrowUpRight, Bookmark } from 'lucide-react';
 import Image from 'next/image';
